requirements.txt
numpy>=1.19.0

networkx>=2.5

pyfastg>=0.1.0

matplotlib>=3.3.0
seaborn>=0.11.0

biopython>=1.78

pandas>=1.1.0
